requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.2.1
colorama==0.4.6
dj-database-url==3.0.0
Django==5.2.3
django-environ==0.12.0
djangorestframework==3.16.0
environ==1.0
gunicorn==23.0.0
h11==0.16.0
packaging==25.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3
whitenoise==6.9.0
